refactor: Remove panics from library code and add typed assembler errors
- fc-rom: replace try_into().unwrap() in format/header parsing with
  infallible byte reads; guard section offset+len against overflow
- fc-asm: introduce EvalError and DirectiveError enums (thiserror),
  replacing stringly-typed Result<T, String> in expr/directive internals
- fc-asm: make Preprocessor::constants/macros private with accessors
- fc-vm: recover from poisoned sound mutex instead of panicking
- fc-hub: fail request with 500 instead of unwrap on missing HubState
- fix clippy warnings across workspace (needless casts, range loops,
  too_many_arguments on publish_cart via PublishArgs struct)

+++ crates/fc-vm/src/isa/operations/camera.rs
@@ -1,8 +1,8 @@
 use crate::vm::{ExecutionContext, VmFault};
 
 pub fn set_camera_position(ctx: &mut ExecutionContext) -> Result<(), VmFault> {
-    let x = ctx.read_register_value()? as u32;
-    let y = ctx.read_register_value()? as u32;
+    let x = ctx.read_register_value()?;
+    let y = ctx.read_register_value()?;
 
     ctx.camera.set_position(x, y);
     Ok(())
